+++ A4/Mesh.hpp
@@ -13,13 +13,13 @@
 class Mesh : public Primitive {
 public:
   Mesh( const std::string& fname );
-  virtual IntersectionInfo checkRayIntersection(const glm::vec4 &ray_origin, const glm::vec4 &ray_dir, double max_t);
+  virtual IntersectionInfo checkRayIntersection(const glm::dvec4 &ray_origin, const glm::dvec4 &ray_dir, double max_t);
 
   ~Mesh() {
     delete boundingBox;
   }
 private:
-  bool computeBGT(const glm::vec4 &ray_origin, const glm::vec4 &ray_dir, Triangle tri, double &beta, double &gamma, double &t);
+  bool computeBGT(const glm::dvec4 &ray_origin, const glm::dvec4 &ray_dir, Triangle tri, double &beta, double &gamma, double &t);
   friend std::ostream& operator<<(std::ostream& out, const Mesh& mesh);
   NonhierBox *boundingBox;
   std::vector<glm::vec3> m_vertices;
